mv init_from_predefined_model to MutableChannelUnit
