+++ mmrazor/models/mutables/mutable_channel/units/sequential_mutable_channel_unit.py
@@ -7,6 +7,7 @@ import torch.nn as nn
 from mmengine import MMLogger
 
 from mmrazor.models.architectures import dynamic_ops
+from mmrazor.models.mutables.mutable_channel import BaseMutableChannel
 from mmrazor.models.utils import make_divisible
 from mmrazor.registry import MODELS
 from ..mutable_channel_container import MutableChannelContainer
@@ -46,6 +47,12 @@ class SequentialMutableChannelUnit(MutableChannelUnit):
         self.divisor = divisor
         self.min_value = min_value
         self.min_ratio = min_ratio
+
+    @classmethod
+    def init_from_mutable_channel(cls, mutable_channel: BaseMutableChannel):
+        unit = super().init_from_mutable_channel(mutable_channel)
+        unit.mutable_channel = mutable_channel
+        return unit
 
     def prepare_for_pruning(self, model: nn.Module):
         """Prepare for pruning, including register mutable channels."""
